chore: add CSRF validation, apply code suggestions
